// mercally-site/src/app/models/tradeoff.ts
export interface Tradeoff {
  slug: string;
  labelEs: string;
  labelEn: string;
  headlineEs: string;
  headlineEn: string;
  aEs: string;
  aEn: string;
  aBodyEs: string;
  aBodyEn: string;
  bEs: string;
  bEn: string;
  bBodyEs: string;
  bBodyEn: string;
  verdictEs: string;
  verdictEn: string;
}